feat: implement tabs navigation structure and screens for home, profile, and vehicle management modules
