+++ apps/mobile/app/(tabs)/financing.tsx
@@ -1,91 +1,72 @@
+import { ArrowRight } from "lucide-react-native";
 import { Pressable, ScrollView, StyleSheet, Text, View } from "react-native";
-import { SafeAreaView } from "react-native-safe-area-context";
-import { ArrowRight, Bell, CarFront, CircleDollarSign, User } from "lucide-react-native";
 
-import Logo from "@/components/Logo";
+import ScreenSectionHeader from "@/components/ui/ScreenSectionHeader";
 import { colors, fonts, fontSize, spacing } from "@/constants/theme";
 
 export default function FinancingScreen() {
   return (
-    <SafeAreaView style={styles.safeArea}>
-      <View style={styles.container}>
-        <ScrollView
-          showsVerticalScrollIndicator={false}
-          contentContainerStyle={styles.contentContainer}
-        >
-          <Logo size={36} />
+    <View style={styles.container}>
+      <ScrollView
+        showsVerticalScrollIndicator={false}
+        contentContainerStyle={styles.contentContainer}
+      >
+        <ScreenSectionHeader
+          title="Financiamento"
+          subtitle="Acompanhe o seu financiamento"
+          style={styles.sectionHeader}
+        />
 
-          <Text style={styles.title}>Financiamento</Text>
-          <Text style={styles.subtitle}>Acompanhe o seu financiamento</Text>
+        <View style={styles.financeCard}>
+          <Text style={styles.carTitle}>Corolla Altis 2026</Text>
 
-          <View style={styles.financeCard}>
-            <Text style={styles.carTitle}>Corolla Altis 2026</Text>
-
-            <View style={styles.installmentRow}>
-              <Text style={styles.installmentLabel}>Parcelas pagas:</Text>
+          <View style={styles.installmentRow}>
+            <Text style={styles.installmentLabel}>Parcelas pagas:</Text>
+            <View style={styles.installmentBadgeWrapper}>
               <View style={styles.installmentBadgeShadow} />
               <View style={styles.installmentBadge}>
                 <Text style={styles.installmentText}>30 / 60</Text>
               </View>
             </View>
-
-            <Text style={styles.bankInfo}>
-              Instituição financeira:{" "}
-              <Text style={styles.bankName}>Banco Toyota do Brasil S.A</Text>
-            </Text>
           </View>
 
-          <View style={styles.invoiceCard}>
-            <Text style={styles.invoiceTitle}>2° Via do boleto</Text>
-            <Text style={styles.invoiceDescription}>
-              Baixe seus boletos e carnês com facilidade
-            </Text>
-
-            <View style={styles.invoiceButtonWrapper}>
-              <View style={styles.invoiceButtonShadow} />
-              <Pressable style={styles.invoiceButton} android_ripple={{ color: "#c70818" }}>
-                <Text style={styles.invoiceButtonText}>Acessar boleto</Text>
-                <ArrowRight size={28} strokeWidth={2.4} color={colors.white} />
-              </Pressable>
-            </View>
-          </View>
-        </ScrollView>
-
-        <View style={styles.navigationBar}>
-          <CircleDollarSign size={22} strokeWidth={1.8} color={colors.black} />
-          <CarFront size={22} strokeWidth={1.8} color={colors.black} />
-          <Bell size={22} strokeWidth={1.8} color={colors.black} />
-          <User size={22} strokeWidth={1.8} color={colors.black} />
+          <Text style={styles.bankInfo}>
+            Instituição financeira:{" "}
+            <Text style={styles.bankName}>Banco Toyota do Brasil S.A</Text>
+          </Text>
         </View>
-      </View>
-    </SafeAreaView>
+
+        <View style={styles.invoiceCard}>
+          <Text style={styles.invoiceTitle}>2° Via do boleto</Text>
+          <Text style={styles.invoiceDescription}>
+            Baixe seus boletos e carnês com facilidade
+          </Text>
+
+          <View style={styles.invoiceButtonWrapper}>
+            <View style={styles.invoiceButtonShadow} />
+            <Pressable style={styles.invoiceButton} android_ripple={{ color: "#c70818" }}>
+              <Text style={styles.invoiceButtonText}>Acessar boleto</Text>
+              <ArrowRight size={28} strokeWidth={2.4} color={colors.white} />
+            </Pressable>
+          </View>
+        </View>
+      </ScrollView>
+    </View>
   );
 }
 
 const styles = StyleSheet.create({
-  safeArea: {
-    flex: 1,
-    backgroundColor: colors.grayLight,
-  },
   container: {
     flex: 1,
+    backgroundColor: colors.background,
   },
   contentContainer: {
-    paddingTop: spacing.md + spacing.xs,
+    paddingTop: spacing.sm,
     paddingHorizontal: spacing.lg + 3,
     paddingBottom: spacing.xxl + spacing.xxl,
   },
-  title: {
-    marginTop: spacing.xxl + spacing.lg + spacing.md,
-    fontFamily: fonts.semiBold,
-    fontSize: fontSize.xl + spacing.xs,
-    color: colors.textPrimary,
-  },
-  subtitle: {
-    marginTop: spacing.xs - 2,
-    fontFamily: fonts.medium,
-    fontSize: fontSize.md,
-    color: "#5C5C5C",
+  sectionHeader: {
+    marginTop: spacing.lg,
   },
   financeCard: {
     marginTop: spacing.xl + spacing.md - 1,
@@ -105,27 +86,31 @@ const styles = StyleSheet.create({
     marginTop: spacing.md - 1,
     flexDirection: "row",
     alignItems: "center",
-    gap: spacing.md,
   },
   installmentLabel: {
     fontFamily: fonts.semiBold,
     fontSize: fontSize.md,
     color: colors.textPrimary,
+    width: 108,
+  },
+  installmentBadgeWrapper: {
+    width: 100,
+    height: 39,
+    marginLeft: spacing.md + 1,
   },
   installmentBadgeShadow: {
     position: "absolute",
-    left: 129,
-    top: 4,
-    width: 100,
-    height: 39,
+    top: 3,
+    left: 3,
+    right: -3,
+    bottom: -3,
     borderWidth: 1,
     borderColor: colors.black,
     backgroundColor: "transparent",
   },
   installmentBadge: {
-    marginLeft: spacing.xl,
-    width: 100,
-    height: 39,
+    width: "100%",
+    height: "100%",
     borderWidth: 1,
     borderColor: colors.black,
     backgroundColor: colors.primary,
@@ -164,7 +149,7 @@ const styles = StyleSheet.create({
     marginTop: spacing.xs - 2,
     fontFamily: fonts.regular,
     fontSize: fontSize.md,
-    color: "#5C5C5C",
+    color: colors.textSecondary,
   },
   invoiceButtonWrapper: {
     width: 219,
@@ -201,23 +186,5 @@ const styles = StyleSheet.create({
     fontFamily: fonts.semiBold,
     fontSize: fontSize.xl,
     color: colors.white,
-  },
-  navigationBar: {
-    position: "absolute",
-    left: 0,
-    right: 0,
-    bottom: 0,
-    height: spacing.xxl + 2,
-    backgroundColor: colors.white,
-    borderTopWidth: 1,
-    borderTopColor: colors.border,
-    flexDirection: "row",
-    alignItems: "center",
-    justifyContent: "space-around",
-    shadowColor: colors.black,
-    shadowOffset: { width: 0, height: -1 },
-    shadowOpacity: 0.12,
-    shadowRadius: 10,
-    elevation: 8,
   },
 });
